// src/users/models/services/generated/user.generated.ts
export interface ErrorResponse {
    statusCode: number;
    message: string;
    timestamp: string;
    path: string;
    "x-request-id": string;
}